requirements.txt
sqlalchemy
psycopg2
Flask-SQLAlchemy
flask_marshmallow
requests==2.23.0